change back to int

+++ backend/models/product.go
@@ -10,7 +10,7 @@ type Product struct {
 	PictureURL    string  `json:"PictureURL"`
 	ProductTitle  string  `json:"ProductTitle"`
 	Price         float32 `json:"Price"`
-	Amount        float32 `json:"Amount"`
+	Amount        int     `json:"Amount"`
 	ProductType   int     `json:"ProductType"`
 	ProductDetail string  `json:"ProductDetail"`
 	Rating        float64 `json:"Rating"`
